+++ CHANGELOG.rst
@@ -2,6 +2,7 @@ Master Branch
 =============
 
 Adding junk.
+Adding junk.
 
 Version 1.0.6 (2020-07-27)
 ==========================